move Leaf to transaction module
